The DelItemCheck() on line 565 was not working.
I do not know if it was changed or if it was moved. (As a sidenote however it seem like it is removed all together in 20.05. I have not looked deeper into 20.05 though as we did not have that need right now.)

The error solved here is:
Undefined subroutine &Koha::Illbackends::Libris::Base::DelItemCheck called at /usr/share/koha/lib/Koha/Illbackends/Libris/Base.pm line 565

And this by adding C4::Items