Update for Koha 22.11, merge waiting branch
Some new features were held on the "waiting" branch, because
they relied on features that were not yet part of a stable
version of Koha. These features are now part of Koha 22.11, and
the changes to the Libris backend that relied on them have been
merged into the "master" branch. This means Koha 22.11 is now
required to run the "master" branch of koha-illbackend-libris.