Implement update_lastseen_for_articlecopies
Enabling "update_lastseen_for_articlecopies" will update the
"lastseen" property of a patron when an article copy is received
for that patron. The update must be connected to an "activity"
listed in the TrackLastPatronActivityTriggers syspref. There is
currently no ILL "activity", so this patch reuses the "article"
activity.

This patch relies on functionality avaiable from Koha 23.11.